+++ test/methods/config.js
@@ -229,9 +229,6 @@ Test('config', function () {
     tx(function () {BigNumber.config({ALPHABET: '0+'})}, "ALPHABET: '0+'");
     tx(function () {BigNumber.config({ALPHABET: '0123456789.'})}, "ALPHABET: '0123456789.'");
 
-    BigNumber.config({ALPHABET: '0,'});
-    t('0,', BigNumber.config().ALPHABET);
-
     BigNumber.config({ALPHABET: 'xy'});
     t('xy', BigNumber.config().ALPHABET);