+++ test/test.js
@@ -33,6 +33,7 @@ console.log('\n Testing bignumber.js\n');
   'toFormat',
   'toFraction',
   'toNumber',
+  'toObject',
   'toPrecision',
   'toString'
 ]
@@ -44,7 +45,6 @@ console.log('\n Testing bignumber.js\n');
   // Reset BigNumber for each method tested?
   //delete require.cache[require.resolve('../bignumber.js')];
   //BigNumber = require('../bignumber');
-  //BigNumber.DEBUG = true;
 });
 
 time = process.hrtime(time);