+++ test/typescript/test_global.ts
@@ -1,0 +1,5 @@
+const n: BigNumber.Value = 3;
+
+const x: BigNumber.Instance = new BigNumber(n);
+
+console.log(x.toString());