+++ test/typescript/test_require.ts
@@ -1,0 +1,7 @@
+import BigNumber = require("bignumber.js");
+
+const n: BigNumber.Value = 3;
+
+const x: BigNumber.Instance = new BigNumber(n);
+
+console.log(x.toString());